Insight-Watt-Backend/pyproject.toml
[project]
name = "insight-watt-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types>=0.7.0",
    "anyio>=4.12.0",
    "beautifulsoup4>=4.14.3",
    "cachetools>=6.2.4",
    "certifi>=2025.11.12",
    "charset-normalizer>=3.4.4",
    "colorama>=0.4.6",
    "contourpy>=1.3.3",
    "cycler>=0.12.1",
    "fastapi>=0.124.4",
    "fonttools>=4.61.1",
    "google>=3.0.0",
    "google-ai-generativelanguage>=0.6.15",
    "google-api-core>=2.25.2",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.45.0",
    "google-auth-httplib2>=0.3.0",
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "googleapis-common-protos>=1.72.0",
    "groq>=1.0.0",
    "grpcio>=1.76.0",
    "grpcio-status>=1.71.2",
    "h11>=0.16.0",
    "httpcore>=1.0.9",
    "httplib2>=0.31.0",
    "httpx>=0.28.1",
    "idna>=3.11",
    "joblib>=1.5.3",
    "jsonpatch>=1.33",
    "jsonpointer>=3.0.0",
    "kiwisolver>=1.4.9",
    "langchain-core>=1.2.5",
    "langgraph>=1.0.5",
    "langgraph-checkpoint>=3.0.1",
    "langgraph-prebuilt>=1.0.5",
    "langgraph-sdk>=0.3.1",
    "langsmith>=0.5.1",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.8",
    "narwhals>=2.14.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.38.0",
]
